# contracts/openapi/work/legacy-web.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Study2Work legacy web API",
    "version": "1.0.0",
    "description": "Compatibility JSON API used by the React Work Web. It maps the current Express/Prisma legacy Work schema and is separate from the target Work domain catalog."
  },
  "servers": [
    { "url": "/api/v1" }
  ],
  "tags": [
    { "name": "Auth" },
    { "name": "Jobs" },
    { "name": "Student" },
    { "name": "Business" }
  ],
  "paths": {
    "/auth/student/login": {
      "post": {
        "tags": ["Auth"],
        "operationId": "loginStudent",
        "requestBody": { "$ref": "#/components/requestBodies/Login" },
        "responses": { "200": { "$ref": "#/components/responses/ApiEnvelope" }, "400": { "$ref": "#/components/responses/ApiError" }, "401": { "$ref": "#/components/responses/ApiError" } }
      }
    },
    "/auth/business/login": {
      "post": {
        "tags": ["Auth"],
        "operationId": "loginBusiness",
        "requestBody": { "$ref": "#/components/requestBodies/Login" },
        "responses": { "200": { "$ref": "#/components/responses/ApiEnvelope" }, "400": { "$ref": "#/components/responses/ApiError" }, "401": { "$ref": "#/components/responses/ApiError" } }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "operationId": "logout",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "$ref": "#/components/responses/ApiEnvelope" }, "401": { "$ref": "#/components/responses/ApiError" } }
      }
    },
    "/students": {
      "post": {
        "tags": ["Student"],
        "operationId": "registerStudent",
        "requestBody": { "$ref": "#/components/requestBodies/StudentRegistration" },
        "responses": { "201": { "$ref": "#/components/responses/ApiEnvelope" }, "400": { "$ref": "#/components/responses/ApiError" }, "409": { "$ref": "#/components/responses/ApiError" } }
      }
    },
    "/jobs": {
      "get": {
        "tags": ["Jobs"],
        "operationId": "listJobs",
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/Limit" }
        ],
        "responses": { "200": { "$ref": "#/components/responses/ApiEnvelope" } }
      }
    },
    "/jobs/{id}": {
      "get": {
        "tags": ["Jobs"],
        "operationId": "getJob",
        "parameters": [ { "$ref": "#/components/parameters/JobId" } ],
        "responses": { "200": { "$ref": "#/components/responses/ApiEnvelope" }, "400": { "$ref": "#/components/responses/ApiError" }, "404": { "$ref": "#/components/responses/ApiError" } }
      }
    },
    "/me": {
      "get": {
        "tags": ["Auth"],
        "operationId": "getMe",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "$ref": "#/components/responses/ApiEnvelope" }, "401": { "$ref": "#/components/responses/ApiError" }, "404": { "$ref": "#/components/responses/ApiError" } }
      }
    },
    "/students/me/cv": {
      "get": {
        "tags": ["Student"],
        "operationId": "getMyCv",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "$ref": "#/components/responses/ApiEnvelope" }, "401": { "$ref": "#/components/responses/ApiError" }, "403": { "$ref": "#/components/responses/ApiError" }, "404": { "$ref": "#/components/responses/ApiError" } }
      },
      "post": {
        "tags": ["Student"],
        "operationId": "createCv",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "$ref": "#/components/requestBodies/CvForm" },
        "responses": { "201": { "$ref": "#/components/responses/ApiEnvelope" }, "400": { "$ref": "#/components/responses/ApiError" }, "401": { "$ref": "#/components/responses/ApiError" }, "403": { "$ref": "#/components/responses/ApiError" }, "409": { "$ref": "#/components/responses/ApiError" } }
      }
    },
    "/students/me/cv/{cvId}": {
      "put": {
        "tags": ["Student"],
        "operationId": "updateCv",
        "security": [{ "bearerAuth": [] }],
        "parameters": [ { "$ref": "#/components/parameters/CvId" } ],
        "requestBody": { "$ref": "#/components/requestBodies/CvForm" },
        "responses": { "200": { "$ref": "#/components/responses/ApiEnvelope" }, "400": { "$ref": "#/components/responses/ApiError" }, "401": { "$ref": "#/components/responses/ApiError" }, "403": { "$ref": "#/components/responses/ApiError" }, "404": { "$ref": "#/components/responses/ApiError" } }
      }
    },
    "/students/me/applications": {
      "get": {
        "tags": ["Student"],
        "operationId": "getMyApplications",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "$ref": "#/components/responses/ApiEnvelope" }, "401": { "$ref": "#/components/responses/ApiError" }, "403": { "$ref": "#/components/responses/ApiError" } }
      }
    },
    "/jobs/{jobId}/applications": {
      "post": {
        "tags": ["Student"],
        "operationId": "applyToJob",
        "security": [{ "bearerAuth": [] }],
        "parameters": [ { "$ref": "#/components/parameters/JobIdPath" } ],
        "responses": { "201": { "$ref": "#/components/responses/ApiEnvelope" }, "400": { "$ref": "#/components/responses/ApiError" }, "401": { "$ref": "#/components/responses/ApiError" }, "403": { "$ref": "#/components/responses/ApiError" }, "404": { "$ref": "#/components/responses/ApiError" }, "409": { "$ref": "#/components/responses/ApiError" } }
      }
    },
    "/businesses/me/jobs": {
      "get": {
        "tags": ["Business"],
        "operationId": "listBusinessJobs",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "$ref": "#/components/responses/ApiEnvelope" }, "401": { "$ref": "#/components/responses/ApiError" }, "403": { "$ref": "#/components/responses/ApiError" } }
      },
      "post": {
        "tags": ["Business"],
        "operationId": "createBusinessJob",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "$ref": "#/components/requestBodies/JobForm" },
        "responses": { "201": { "$ref": "#/components/responses/ApiEnvelope" }, "400": { "$ref": "#/components/responses/ApiError" }, "401": { "$ref": "#/components/responses/ApiError" }, "403": { "$ref": "#/components/responses/ApiError" } }
      }
    },
    "/businesses/me/jobs/{jobId}": {
      "put": {
        "tags": ["Business"],
        "operationId": "updateBusinessJob",
        "security": [{ "bearerAuth": [] }],
        "parameters": [ { "$ref": "#/components/parameters/JobIdPath" } ],
        "requestBody": { "$ref": "#/components/requestBodies/JobForm" },
        "responses": { "200": { "$ref": "#/components/responses/ApiEnvelope" }, "400": { "$ref": "#/components/responses/ApiError" }, "401": { "$ref": "#/components/responses/ApiError" }, "403": { "$ref": "#/components/responses/ApiError" }, "404": { "$ref": "#/components/responses/ApiError" } }
      },
      "delete": {
        "tags": ["Business"],
        "operationId": "deleteBusinessJob",
        "security": [{ "bearerAuth": [] }],
        "parameters": [ { "$ref": "#/components/parameters/JobIdPath" } ],
        "responses": { "200": { "$ref": "#/components/responses/ApiEnvelope" }, "401": { "$ref": "#/components/responses/ApiError" }, "403": { "$ref": "#/components/responses/ApiError" }, "404": { "$ref": "#/components/responses/ApiError" } }
      }
    },
    "/businesses/me/applications": {
      "get": {
        "tags": ["Business"],
        "operationId": "getBusinessApplications",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "$ref": "#/components/responses/ApiEnvelope" }, "401": { "$ref": "#/components/responses/ApiError" }, "403": { "$ref": "#/components/responses/ApiError" } }
      }
    },
    "/students/{studentId}/cv": {
      "get": {
        "tags": ["Business"],
        "operationId": "getStudentCv",
        "security": [{ "bearerAuth": [] }],
        "parameters": [ { "$ref": "#/components/parameters/StudentId" } ],
        "responses": { "200": { "$ref": "#/components/responses/ApiEnvelope" }, "400": { "$ref": "#/components/responses/ApiError" }, "401": { "$ref": "#/components/responses/ApiError" }, "403": { "$ref": "#/components/responses/ApiError" }, "404": { "$ref": "#/components/responses/ApiError" } }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "parameters": {
      "Page": { "name": "page", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "default": 1 } },
      "Limit": { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 50, "default": 6 } },
      "JobId": { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64", "minimum": 1 } },
      "JobIdPath": { "name": "jobId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64", "minimum": 1 } },
      "CvId": { "name": "cvId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64", "minimum": 1 } },
      "StudentId": { "name": "studentId", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64", "minimum": 1 } }
    },
    "requestBodies": {
      "Login": {
        "required": true,
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } } }
      },
      "StudentRegistration": {
        "required": true,
        "content": { "multipart/form-data": { "schema": { "$ref": "#/components/schemas/StudentRegistration" } } }
      },
      "CvForm": {
        "required": true,
        "content": { "multipart/form-data": { "schema": { "$ref": "#/components/schemas/CvForm" } } }
      },
      "JobForm": {
        "required": true,
        "content": { "multipart/form-data": { "schema": { "$ref": "#/components/schemas/JobForm" } } }
      }
    },
    "responses": {
      "ApiEnvelope": {
        "description": "Successful response envelope.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiEnvelope" } } }
      },
      "ApiError": {
        "description": "Error response envelope.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiEnvelope" } } }
      }
    },
    "schemas": {
      "ApiEnvelope": {
        "type": "object",
        "required": ["success", "businessCode", "message", "data", "meta", "traceId"],
        "properties": {
          "success": { "type": "boolean" },
          "businessCode": { "type": "string" },
          "message": { "type": "string" },
          "data": {},
          "meta": { "type": "object", "additionalProperties": true },
          "traceId": { "type": "string" }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": { "email": { "type": "string", "format": "email" }, "password": { "type": "string", "minLength": 1 } }
      },
      "StudentRegistration": {
        "type": "object",
        "required": ["hoten", "email", "matkhau"],
        "properties": {
          "hoten": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "matkhau": { "type": "string" },
          "chuyennganh": { "type": "string" },
          "avt": { "type": "string", "format": "binary" }
        }
      },
      "CvForm": {
        "type": "object",
        "required": ["hoten", "email"],
        "properties": {
          "hoten": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "ngaysinh": { "type": "string", "format": "date" },
          "gioitinh": { "type": "string" },
          "sdt": { "type": "string" },
          "diachi": { "type": "string" },
          "vitri": { "type": "string" },
          "nganh": { "type": "string" },
          "muctieunghiep": { "type": "string" },
          "hocvan": { "type": "string" },
          "kinhnghiem": { "type": "string" },
          "kynang": { "type": "string" },
          "ngoaingu": { "type": "string" },
          "chungchi": { "type": "string" },
          "duan": { "type": "string" },
          "giaithuong": { "type": "string" },
          "hoatdong": { "type": "string" },
          "social": { "type": "string" },
          "portfolio": { "type": "string" },
          "luongmongmuon": { "type": "string" },
          "avt": { "type": "string", "format": "binary" }
        }
      },
      "JobForm": {
        "type": "object",
        "required": ["ten_vi_tri", "dia_diem"],
        "properties": {
          "ten_vi_tri": { "type": "string" },
          "phong_ban": { "type": "string" },
          "cap_bac": { "type": "string" },
          "bao_cao_cho": { "type": "string" },
          "nhiem_vu": { "type": "string" },
          "trinh_do": { "type": "string" },
          "kinh_nghiem": { "type": "string" },
          "ky_nang": { "type": "string" },
          "ky_nang_mem": { "type": "string" },
          "uu_tien": { "type": "string" },
          "muc_luong": { "type": "string" },
          "phuc_loi": { "type": "string" },
          "moi_truong": { "type": "string" },
          "dia_diem": { "type": "string" },
          "thoi_gian": { "type": "string" },
          "han_nop": { "type": "string" },
          "cach_ung_tuyen": { "type": "string" },
          "mo_ta": { "type": "string" },
          "ten_cong_ty": { "type": "string" },
          "nganh": { "type": "string" },
          "avt": { "type": "string", "format": "binary" }
        }
      }
    }
  }
}